refactor: 기존의 useProducts, useCartItems를 useProductsQuery, useCartItemQuery로 대체
